move ChannelSettings to base acquisition settings
